Refactor common code between the significance testers
This moves stats and functionality up towards parent classes
so that generic stats like mean are always computed for
all variants. In fact TTestStat is now an empty class,
it's functionality fully absorbed by the parent. (But note
that the class name/type itself carries information about the pvalue.
